refactor - provideWinnerNames function is added in Winners class and modify tests
