+++ src/main/java/racingcar/Runner.java
@@ -1,4 +1,6 @@
-package racingcar.controller;
+package racingcar;
+
+import racingcar.controller.RacingGame;
 
 public class Runner {
     public static void main(String[] args) {
