+++ src/main/java/racingcar/constants/ComparableConstants.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.constants;
 
 public class ComparableConstants {
     public static final int GREATER_THAN = 1;
