+++ src/main/java/racingcar/view/ConsoleOutputView.java
@@ -48,14 +48,8 @@ public class ConsoleOutputView implements OutputView{
     public void showWinners(Winners winners) {
         StringBuilder stringBuilder = new StringBuilder();
 
-        winners.provideWinnerNames().forEach( name -> {
-            stringBuilder.append(name);
-            stringBuilder.append(SEPARATOR_OF_WINNERS);
-        });
-
-        int lengthOfMessage = stringBuilder.length();
-        stringBuilder.delete(lengthOfMessage-SEPARATOR_OF_WINNERS.length(), lengthOfMessage);
-
+        String winnerNames = winners.provideWinnerNames();
+        stringBuilder.append(winnerNames);
         stringBuilder.append(MESSAGE_OF_WINNERS);
 
         System.out.println(stringBuilder);
